requirements.txt
customtkinter==5.2.0
pymongo==4.6.0
openai==1.93.3
transformers==4.36.0
pylint==3.0.0
flake8==6.1.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
chardet==5.2.0
setuptools>=65.0.0  # DODANE - zamiennik dla distutils
wheel>=0.37.0       # DODANE - wspiera setuptools
dnspython>=2.3.0    # DODANE - dla MongoDB

packaging>=21.0
darkdetect>=0.8.0
cryptography>=41.0.0   